requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
openai>=1.54.0
anthropic>=0.39.0
pyairtable>=3.0.0
redis>=5.0.0
pydantic>=2.9.0
tenacity>=9.0.0
slack-bolt>=1.22.0
slack-sdk>=3.35.0
aiohttp>=3.11.0
apscheduler>=3.11.0
pytz>=2024.1
